uplink/eestream: fixes to decode logic
this changes the conditions at which we attempt to issue decode
calls to the underlying codec. this now attempts to decode
whenever we have enough shares and continues to do so when
new shares are added until the decode succeeds or we fail and
do not have any possible shares to return.

the old logic when force error detection was false (which is
apparently always) was not correct when calculating if
enough shares were reached, which meant that we de-facto had
error detection on regardless.

the random decode tests were flaky: with only one extra share
you can only be certain to detect 1 error, so any tests with
more than 1 error could successfully decode but fail the
check that the data was the same.

the slow tests were insufficient due to the mentioned problem
with error detection. for example, if you have 2 required out
of 4 total and 2 of the total nodes are slow, you should be
able to reconstruct the file without delay. the old code could
not.

since we actually do want error detection on in prod, we add
a configuration field that by default turns it on, so that
there should not be much observable difference.

Change-Id: I97700f5e47c2ad5effebec70d6a545c2fc459eee