satellite/gracefulexit: make disabled test more permissive
the test required that the error have rpc status of Unimplemented
but the code does not depend on this detail. drpc does not have
that specific error code, nor a clean way to tell that the error
is due to it being an unknown rpc. so, just assert that it got
an error.

Change-Id: I9e54a8b5297ec1cfc5f7206bca1f84c76ce257bf