satellite/repair: new timeout applying to entire repair op
There are already config items for Timeout and DownloadTimeout, which
govern long-tail cutoff times for uploading repaired pieces and
downloading pieces for repair, respectively. But we need something
that covers the entire repair operation, including the point in time
where the repair entry is removed from the queue, so we can safely
expect that a repairer will not work on an entry from the queue for
more than this time period.

This still isn't 100% reliable: if the host is under load, the context
timeout goroutine might not be scheduled immediately when the timeout
has elapsed, and it may take an unspecified amount of time for the
canceler to run and for all of the relevant pieces to quit. We allow
45 minutes for that (15 minutes HardRepairTimeout + 45 minutes = 1 hour
time before other repair workers are allowed to pick up an entry on the
repair queue showing a recent attempt time).

We should fix this at some point so that repairs are more safely kept to
a single worker but also returned safely to the queue if that worker
crashes.

Change-Id: I4db1a6160329d4006cce68e8a625ff442b9eef85