REVIEW: clean up remote & inline segments
We have to remove all the segments which have been uploaded to the
satellite which means to delete remote and inline segments after the
satellite request for them has succeeded.

Current segment is used for clean up, so it has to be incremented only
when the request to the satellite for uploading a new segment has
succeeded.