cmd/segment-reaper: Remove Cluster type
Cluster struct type is used as a key for a ObjectsMap type.
Observer struct type has a field of ObjectsMap.
Cluster has a field for project ID.
Observer processSegment method uses its ObjectMap field for tracking
objects.

However Observer processSegment clears the map once the projectID
diverges from the one kept in the Observer lastProjectID field, meaning
that it isn't needed to keep the projectID as part of the ObjectMap key.

For this reason, ObjectMap can use as a key just only the bucket name
and Cluster struct isn't needed.

Because of such change the ObjectMap type has been renamed to a more
descriptive name.