cmd/segment-reaper: Make test values more random in test
Using math/rand package generates a deterministic values and without
seeding them the values are the same on each test run on the same
machine while it's not rebooted.

Seeding the values in the test Main using the unix timestamp provokes
that the random values varies between different test runs, giving the
chance to detect test failures during the development cycle.