storage/redis: Limit should not be applied as count.
COUNT on a SCAN does not actually limit the results [1]. It limits the
amount of work a single call to SCAN will perform before returning. By
setting this to limit we can sometimes timeout on the request if limit
is very large.

This restores storage/redis back to it's original behavior.

[1]: https://redis.io/commands/scan#the-count-option

Change-Id: Ia75afb5152df909df38c9a7c6feb74d062f49d6a